requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.8.0rc2
Deprecated==1.2.18
diff-match-patch==20241021
Django==4.2.18
django-autoslug==1.9.8
django-cors-headers==3.11.0
django-crontab==0.7.1
django-decouple==2.1
django-filter==21.1
django-import-export==4.3.5
django-jazzmin==2.6.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.0.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
phonenumbers==8.13.55
Pillow==9.1.0
prompt_toolkit==3.0.50
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==4.2.2
requests==2.27.1
requests-oauthlib==2.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
XlsxWriter==3.2.2
